Work 3/31
** Changes **
- UserHandlerService: Removed checkExists

** Issues **
- Still working on the async call to the firebase database.  To check if
the user exists, then act depending on the results.  Currently the calls
to firebase for information
(this.angularfire.database.object(path).$ref.once) is an asyc call and
if it is encompassed in a function, the function will return a value
before the call is complete.

Currently unsure how to make the function wait or reorganize the call
differently to encompass the synced results to work within the confines
of the async call