Work 4/3
4/3:
** Changes **
- UserHandlerService: Figured out how to successfully check that a user
exists in the UserHandlerService.  The solution was to create a custom
promise within the checkExists method.  This allowed me to add a then
and catch promise to the method call elsewhere in the program.
- Example: App-Enter is able to call checkExists during the Login method
and use the then promise to wait for a resolution from checkExists.
Then it can use the same login process as was programmed into it.
- UserHandlerService: Added Documenation and Comments
- App-Enter: Updated login method to work through UserHandlerService
- Entry Component: Updated getUser() as much of the changes that
happened are now happening in the UserHandlerService
- User-Tile: Updated ngOnChanges because it is no longer neccessary to
push the users name from app-enter > entry > user-tile

** Issues **
- [] Updates to User-Tile caused an error message when the page loads as
the page loading triggers ngOnChanges